Moved auth logic to another file
